requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic

langchain
langchain-community
langchain-text-splitters

chromadb
sentence-transformers

google-generativeai
protobuf

scikit-learn
numpy
shap

neo4j
langgraph
pypdf
